[#5] fix: MemberReadServiceTest 삭제
MemberReadServiceTest 삭제로 인한 테스트 케이스 변경
